Added links between the views.
